+++ src/pages/HomePages.js
@@ -2,24 +2,13 @@ import React, { useEffect, useState } from 'react';
 
 import Header from 'components/Header/Header';
 import { Outlet } from 'react-router-dom';
-import { TestThemeBtn } from 'components/TestThemeBtn/TestThemeBtn';
+
 // import ScreensPage from 'components/ScreensPage/ScreensPage';
 // import Sidebar from 'components/Sidebar/Sidebar';
 // import Backdrop from '../components/Sidebar/Backdrop';
 
 const HomePage = () => {
   const [showSidebar, setShowSidebar] = useState(true);
-  const [theme, setTheme] = useState('dark');
-
-  useEffect(() => {
-    document.body.classList.add(theme);
-    return () => {
-      document.body.classList.remove(theme);
-    };
-  }, [theme]);
-  const handleThemeChange = color => {
-    setTheme(color);
-  };
 
   const toggleSidebar = () => {
     setShowSidebar(!showSidebar);
@@ -56,20 +45,6 @@ const HomePage = () => {
           }}
         >
           <Header onToggleSidebar={toggleSidebar} />
-          <TestThemeBtn
-            onClick={() => {
-              handleThemeChange('violet');
-            }}
-          >
-            Тема violet
-          </TestThemeBtn>
-          <TestThemeBtn
-            onClick={() => {
-              handleThemeChange('dark');
-            }}
-          >
-            Тема dark
-          </TestThemeBtn>
           <Outlet />
           {/* <ScreensPage /> */}
         </div>
